Loading popups, unlink spotify abiliity

+++ partyfy/pages/api/database/users.ts
@@ -31,6 +31,12 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse<
     }
     // Delete a user's profile
     if (req.method === 'DELETE') {
+        let action = req.query.action as string;
+        if (action === 'unlink') {
+            let data = await database.unlinkUser(req.query.UserID as string) as any;
+            res.status(200).json(data);
+            return;
+        }
         let data = await database.deleteUser(req.query.UserID as string) as any;
         res.status(200).json(data);
         return;
